add day condition

+++ app/src/main/java/com/example/weatherapp/Model/WeatherDataClass.kt
@@ -2,7 +2,7 @@ import com.google.gson.annotations.SerializedName
 
 data class WeatherDataClass (
 
-	@SerializedName("now") val now : Int,
+	@SerializedName("now") val now : Long,
 	@SerializedName("now_dt") val now_dt : String,
 	@SerializedName("info") val info : Info,
 	@SerializedName("fact") val fact : Fact,
